api/requirements.txt
CherryPy==18.8.0
PyYAML==6.0